Moving back and to more minimal solutions
I think that this project does not need CMake, which is a large piece of
softare that is useful in cases where the code needs to run on different
environments. In this case, this code will only run in UNIX-like systems
using X11 or through Cygwin which is similar to UNIX-like environment.

I have reversed 8a8f333cfe8db91a5d49941b8670ed0dfd61df30.

I have moved the documentation from asciidoc to mdoc, removing
the need for conversion.